save and extract pivot options #5
